Move get function token public key to _token file
